Refactor state
