+++ EasyReflectometryApp/Gui/Pages/Summary/MainContentReport.qml
@@ -134,7 +134,7 @@ Item {
 
     onHtmlChanged: {
         //print(html)
-        ExGlobals.Constants.proxy.setReport(html)
+        ExGlobals.Constants.proxy.project.setReport(html)
         webView.loadHtml(html)
     }
 
@@ -472,7 +472,7 @@ Item {
     property string fittingInfo: {
         if (!isFitting)
             return ''
-        const redchi2 = ExGlobals.Constants.proxy.fitResults.redchi2.toFixed(2)
+        const redchi2 = ExGlobals.Constants.proxy.fitter.fitResults.redchi2.toFixed(2)
         let list = [
                 '<p>',
                 `<b>Goodness-of-fit (reduced \u03c7\u00b2):</b> ${redchi2}<br>`,
